refactor(workboard): split gateway dispatch helpers
